+++ packages/icrc-ledger-types/CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The `Value` type and the algorithm to compute its hash.
+- The blocks and transactions types for an icrc ledger.